# docs/source/API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reachy Mini REST API",
    "version": "0.1.0",
    "description": "HTTP and WebSocket API exposed by the Reachy Mini daemon. Interactive Swagger docs are available at `/docs` when the daemon is running."
  },
  "paths": {
...
    "/api/hf-auth/save-token": {
      "post": {
        "summary": "Save Token",
        "description": "Save HuggingFace token after validation.",
        "operationId": "save_token_api_hf_auth_save_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hf-auth/oauth/start": {
      "get": {
        "summary": "Start Oauth",
        "description": "Start a new OAuth authorization session.\n\nReturns the auth_url to redirect the user to HuggingFace.\n\nArgs:\n    request: The incoming HTTP request.\n    use_localhost: When True, use localhost:8000 as the OAuth callback URL.\n        Passed by the desktop app which proxies localhost:8000 to the robot.",
        "operationId": "start_oauth_api_hf_auth_oauth_start_get",
        "parameters": [
          {
            "name": "use_localhost",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Use Localhost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Oauth Api Hf Auth Oauth Start Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hf-auth/oauth/status/{session_id}": {
      "get": {
        "summary": "Get Oauth Status",
        "description": "Poll for OAuth session status.\n\nThe frontend polls this endpoint to check if the user has\ncompleted authorization.",
        "operationId": "get_oauth_status_api_hf_auth_oauth_status__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Oauth Status Api Hf Auth Oauth Status  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hf-auth/oauth/session/{session_id}": {
      "delete": {
        "summary": "Cancel Oauth Session",
        "description": "Cancel an OAuth session.",
        "operationId": "cancel_oauth_session_api_hf_auth_oauth_session__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Oauth Session Api Hf Auth Oauth Session  Session Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hf-auth/oauth/device/status/{session_id}": {
      "get": {
        "summary": "Get Device Oauth Status",
        "description": "Poll for device-code login completion.\n\nOn the first poll that reports success, bring the central relay up (a\ntoken-less boot has no relay instance yet), mirroring the OAuth callback.",
        "operationId": "get_device_oauth_status_api_hf_auth_oauth_device_status__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Device Oauth Status Api Hf Auth Oauth Device Status  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hf-auth/oauth/device/session/{session_id}": {
      "delete": {
        "summary": "Cancel Device Oauth Session",
        "description": "Cancel a pending device-code login session.",
        "operationId": "cancel_device_oauth_session_api_hf_auth_oauth_device_session__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Device Oauth Session Api Hf Auth Oauth Device Session  Session Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hf-auth/oauth/callback": {
      "get": {
        "summary": "Oauth Callback",
        "description": "Handle OAuth callback from HuggingFace.\n\nThis is where HF redirects after user authorizes.\nShows a success/error page that the user can close.",
        "operationId": "oauth_callback_api_hf_auth_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "TokenRequest",
        "description": "Request model for saving a HuggingFace token."
      },
      "TokenResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TokenResponse",
        "description": "Response model for token operations."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}